Configure on Message
Change of configuration method: Instead of trying to configure endevices in regular intervals, an attempt to configure a device is made when a message is recieved. The system ensures that only one attempt is made even if several messages are recieved in short order.
If an enddevice cannot be configured after the 10th attempt, the frequency of configure attempts is lowered from every message to every 10th message. After 20 failed attempts, the Adapter will stop trying to configure this enddevice.