A lot of changes related to `composite` exposes and exposes access ...
On branch exposes_composite_enchancements

Changed preparation of the `composite` exposes, to process or create
an objects in getter and setter respectively.
Added creation states, when read part is described in `usual`
expose (`binary` for example) and write part is described in the
`composite` expose.
To make it possible chnanges the push metod in `states` arrey and
additional properties `readable` and `writable` are stored in states.

 Changes to be committed:
	modified:   lib/exposes.js