Groups and Persistance
Device name Persistance:
Use a separate JSON file to retain user modified device and group names in case of adapter reinstall or object deletion. A file dev_names.json will be written to (and read from) iobroker-data/zigbee.x. It contains a mapping of ieee to device name. Device names edited within the zigbee Adapter will be renamed in this file. Note that changes made directly in the object tree are not tracked and will not be retained.

Groups rework, part 2:
- disconnect "device.groups" state
- allow "device.groups" state to be removed via stateCleanup
- disconnect "info.groups" state
- transfer content from "info.groups" to dev_names.json, making "info.groups" obsolete.

Note: this commit does include additional code in admin.js and index_m.html to offer a better group membership overview, but i was unable to get it to run properly, so i have deactivated said code for the time being. This is work in Progress still.